Change behavior of getNormalizedLemma in DictionaryItem. Noe it only removes "-mek -mak" from verb lemmas.
Change behavior of two methods and add two methods in WordAnalysisSurfaceFormatter. Now if apostrophe is provided, it will use it even in regular words. New methods are for convenience. Also when formatting, it will not use lemma directly. This was failing for verbs.

Spellchecker will not throw if a language model is not provided.